AR review. Check ARIMA mods and look for time series signals in other models. New CI.
Residual autocorrelation was detected for all ARIMA models, sibling models and ricker models. auto.arima suggested differencing in all cases which was done. I'm not sure if the sibling or Ricker models retain their interpretation when using differences. Also renamed objects to make clearer what was being done and recorded the code so there was a common patterns followed for each age class. Calculate the prediction interval from past forecasts so that our error included a variability from the model selection process. Still need to work on markdown to add some text and/or show code so that the outputs are understandable to a new reader.