Add Anchor 2025
Still in draft form. Not sure if model averaging is appropriate here bc it does have management implications... i.e. Mike uses 3 year rolling means which forecasts below the lower bound. If the 4/5 sibling holds anywhere close to what it has (at least .9) we will be within the goal. Model averaging includes the sibling for 5 years olds and reflects that.